Add --nocreate-{pg,redis} and --nodrop-{pg,redis} opts in run_test.sh

+++ test/prepare_db.sh
@@ -1,14 +1,31 @@
 #!/bin/sh
-#!/bin/sh
 
 # this script prepare database and redis instance to run acceptance test
 #
 # NOTE: assumes existance of a "template_postgis" loaded with
 #       compatible version of postgis (legacy.sql included)
 
+PREPARE_REDIS=yes
+PREPARE_PGSQL=yes
+
+while [ -n "$1" ]; do
+  if test "$1" = "--skip-pg"; then
+    PREPARE_PGSQL=no
+    shift; continue
+  elif test "$1" = "--skip-redis"; then
+    PREPARE_REDIS=no
+    shift; continue
+  fi
+done
+
+die() {
+        msg=$1
+        echo "${msg}" >&2
+        exit 1
+}
+
 # This is where postgresql connection parameters are read from
 TESTENV=../config/environments/test.js
-
 
 # Extract postgres configuration
 PGHOST=`node -e "console.log(require('${TESTENV}').db_host || '')"`
@@ -43,48 +60,61 @@ if test -z "$TEST_DB"; then
 fi
 TEST_DB=`echo ${TEST_DB} | sed "s/<%= user_id %>/${TESTUSERID}/"`
 
-REDIS_PORT=`node -e "console.log(require('${TESTENV}').redis_port || '6333')"`
-
 export PGHOST PGPORT
 
-die() {
-        msg=$1
-        echo "${msg}" >&2
-        exit 1
-}
+if test x"$PREPARE_PGSQL" = xyes; then
 
-echo "preparing postgres..."
-dropdb ${TEST_DB} # 2> /dev/null # error expected if doesn't exist, but not otherwise
-createdb -Ttemplate_postgis -EUTF8 ${TEST_DB} || die "Could not create test database"
-cat test.sql |
-  sed "s/:PUBLICUSER/${PUBLICUSER}/" |
-  sed "s/:PUBLICPASS/${PUBLICPASS}/" |
-  sed "s/:TESTUSER/${TESTUSER}/" |
-  sed "s/:TESTPASS/${TESTPASS}/" |
-  psql ${TEST_DB} 
+  echo "preparing postgres..."
+  dropdb ${TEST_DB} # 2> /dev/null # error expected if doesn't exist, but not otherwise
+  createdb -Ttemplate_postgis -EUTF8 ${TEST_DB} || die "Could not create test database"
+  cat test.sql |
+    sed "s/:PUBLICUSER/${PUBLICUSER}/" |
+    sed "s/:PUBLICPASS/${PUBLICPASS}/" |
+    sed "s/:TESTUSER/${TESTUSER}/" |
+    sed "s/:TESTPASS/${TESTPASS}/" |
+    psql ${TEST_DB} 
 
-psql -f support/CDB_QueryStatements.sql ${TEST_DB} 
-psql -f support/CDB_QueryTables.sql ${TEST_DB} 
+  # TODO: send in a single run, togheter with test.sql
+  psql -f support/CDB_QueryStatements.sql ${TEST_DB} 
+  psql -f support/CDB_QueryTables.sql ${TEST_DB} 
 
-echo "preparing redis..."
-echo "HSET rails:users:vizzuality id 1" | redis-cli -p ${REDIS_PORT} -n 5
-echo "HSET rails:users:vizzuality database_name ${TEST_DB}" | redis-cli -p ${REDIS_PORT} -n 5
-echo "HSET rails:users:vizzuality" "map_key" "1234" | redis-cli -p ${REDIS_PORT} -n 5
-echo "SADD rails:users:vizzuality:map_key 1235" | redis-cli -p ${REDIS_PORT} -n 5
+fi
 
-# A user configured as with cartodb-2.5.0+
-echo "HSET rails:users:cartodb250user id ${TESTUSERID}" | redis-cli -p ${REDIS_PORT} -n 5
-echo 'HSET rails:users:cartodb250user database_name "'${TEST_DB}'"' | redis-cli -p ${REDIS_PORT} -n 5
-echo 'HSET rails:users:cartodb250user database_host "localhost"' | redis-cli -p ${REDIS_PORT} -n 5
-echo 'HSET rails:users:cartodb250user database_password "'${TESTPASS}'"' | redis-cli -p ${REDIS_PORT} -n 5
-echo "HSET rails:users:cartodb250user map_key 1234" | redis-cli -p ${REDIS_PORT} -n 5
+if test x"$PREPARE_REDIS" = xyes; then
 
-echo "hset rails:oauth_access_tokens:l0lPbtP68ao8NfStCiA3V3neqfM03JKhToxhUQTR consumer_key fZeNGv5iYayvItgDYHUbot1Ukb5rVyX6QAg8GaY2" | redis-cli -p ${REDIS_PORT} -n 3
-echo "hset rails:oauth_access_tokens:l0lPbtP68ao8NfStCiA3V3neqfM03JKhToxhUQTR consumer_secret IBLCvPEefxbIiGZhGlakYV4eM8AbVSwsHxwEYpzx" | redis-cli -p ${REDIS_PORT} -n 3
-echo "hset rails:oauth_access_tokens:l0lPbtP68ao8NfStCiA3V3neqfM03JKhToxhUQTR access_token_token l0lPbtP68ao8NfStCiA3V3neqfM03JKhToxhUQTR" | redis-cli -p ${REDIS_PORT} -n 3
-echo "hset rails:oauth_access_tokens:l0lPbtP68ao8NfStCiA3V3neqfM03JKhToxhUQTR access_token_secret 22zBIek567fMDEebzfnSdGe8peMFVFqAreOENaDK" | redis-cli -p ${REDIS_PORT} -n 3
-echo "hset rails:oauth_access_tokens:l0lPbtP68ao8NfStCiA3V3neqfM03JKhToxhUQTR user_id 1" | redis-cli -p ${REDIS_PORT} -n 3
-echo "hset rails:oauth_access_tokens:l0lPbtP68ao8NfStCiA3V3neqfM03JKhToxhUQTR time sometime" | redis-cli -p ${REDIS_PORT} -n 3
+  REDIS_PORT=`node -e "console.log(require('${TESTENV}').redis_port || '6336')"`
+
+  echo "preparing redis..."
+
+  cat <<EOF | redis-cli -p ${REDIS_PORT} -n 5
+HMSET rails:users:vizzuality \
+ id 1 \
+ database_name ${TEST_DB} \
+ map_key 1234 
+SADD rails:users:vizzuality:map_key 1235
+EOF
+
+  # A user configured as with cartodb-2.5.0+
+  cat <<EOF | redis-cli -p ${REDIS_PORT} -n 5
+HMSET rails:users:cartodb250user \
+ id ${TESTUSERID} \
+ database_name ${TEST_DB} \
+ database_host localhost \
+ database_password ${TESTPASS} \
+ map_key 1234
+EOF
+
+  cat <<EOF | redis-cli -p ${REDIS_PORT} -n 3
+HMSET rails:oauth_access_tokens:l0lPbtP68ao8NfStCiA3V3neqfM03JKhToxhUQTR \
+ consumer_key fZeNGv5iYayvItgDYHUbot1Ukb5rVyX6QAg8GaY2 \
+ consumer_secret IBLCvPEefxbIiGZhGlakYV4eM8AbVSwsHxwEYpzx \
+ access_token_token l0lPbtP68ao8NfStCiA3V3neqfM03JKhToxhUQTR \
+ access_token_secret 22zBIek567fMDEebzfnSdGe8peMFVFqAreOENaDK \
+ user_id 1 \
+ time sometime
+EOF
+
+fi
